LayoutTests/media/modern-media-controls/media-controller/media-controller-fade-controls-when-entering-fullscreen-expected.txt

Fading in the controls when entering fullscreen.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


We should not fade the controls in when showing them inline the first time
PASS shadowRoot.lastChild.classList.contains('fade-in') is false
Clicking on the fullscreen button
Obtained a webkitfullscreenchange event
PASS media.webkitDisplayingFullscreen is true
We should fade the controls in when showing them fullscreen
PASS shadowRoot.lastChild.classList.contains('fade-in') is true

PASS successfullyParsed is true

TEST COMPLETE
